+++ frontend/src/hooks/api/useGetChefs.api.ts
@@ -1,4 +1,4 @@
-import { type GetAllChefsRes } from "../../../../shared/http-types/chef/getAllChefs.http-type";
+import { type GetAllChefsRes } from "../../../../shared/http-types/chef/getAllChefs.api";
 import { useQuery } from "@tanstack/react-query";
 import api from "../../api";
