+++ shared/api/generic/uuidReq.api.ts
