+++ shared/api/recipe/getRecipeByUuid.http-type.ts
@@ -1,5 +1,5 @@
 import { Recipe } from "../../types/recipe.type";
-import { UuidReq } from "../generic/uuidReq.https-types";
+import { UuidReq } from "../generic/uuidReq.api";
 
 export type GetRecipeByIdRes = {
   recipe: Recipe;
